Move forEachDocumentBatchInCollection to new file
